chore: update Rust toolchain to 1.52.1 (#1005)
This branch updates the Rust toolchain to 1.52.1. This includes a [major
bugfix][1] for an issue effecting incremental compilation. Additionally,
Rust 1.51 enabled the `resolver = "2"` feature in Cargo.toml, which can
be used to fix the feature flagging issues with tracing in a more
principled way than the current solution.

A *very* large amount of new lints were added since Rust 1.49.0; in
particular:

* clippy now warns when an `Into` impl could be a `From` impl, because
  `From` impls provide `Into` impls for free, but the reverse is not
  the case
* panic messages must now *always* be `format_args!`
* exact comparisons of floating-point numbers (rather than within an
  error margin) now produce a warning

This branch also fixes all the new warnings.

[1]: https://blog.rust-lang.org/2021/05/10/Rust-1.52.1.html

Signed-off-by: Eliza Weisman <eliza@buoyant.io>